tech13_garage/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
chartjs==1.2
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
distlib==0.3.9
distro==1.9.0
filelock==3.16.1
flake8==7.1.1
Flask==3.0.3
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.0.0
inflect==7.3.1
itsdangerous==2.2.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
Jinja2==3.1.4
jiter==0.8.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.3
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.3.0
mysqlclient==2.2.6
numpy==2.1.3
openai==1.57.1
OpenAIAuth==3.0.0
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
php==1.2.1
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
pycodestyle==2.12.1
pydantic==2.10.3
pydantic_core==2.27.1
pyflakes==3.2.0
Pygments==2.18.0
PyMySQL==1.1.1
pynput==1.7.7
pyodbc==5.2.0
pyparsing==3.2.0
PyQRCode==1.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
pyzbar==0.1.9
qrcode==8.0
regex==2024.11.6
requests==2.32.3
revChatGPT==6.8.6
rich==13.9.4
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.36
tiktoken==0.8.0
tls-client==1.0.1
tomli==2.0.1
tqdm==4.67.1
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.27.1
wcwidth==0.2.13
Werkzeug==3.1.1
wheel==0.43.0
zipp==3.19.2
